auth_service/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.24.0
motor==3.6.1

pydantic[email]==2.5.2
passlib==1.7.4

bcrypt==4.0.1
PyJWT==2.8.0
